Bug fix for nav cookie

+++ src/components/NavBar/NavBar.jsx
@@ -5,7 +5,7 @@ import './navbar.css';
 import { useAuth } from '../../context/auth/authContext';
 import { useNavigate } from 'react-router-dom';
 import { Link } from 'react-router-dom';
-import React from "react";
+import React, { useEffect, useState } from "react";
 import axios from 'axios';
 import { jwtDecode } from 'jwt-decode';
 
@@ -14,10 +14,19 @@ function NavBar() {
     const nav = useNavigate();
     // Destructure logout, isAuth, and toggleAuth from useAuth
     const { logout, isAuth, toggleAuth, cookies } = useAuth();
-    // Decode token from cookies
-    const decoded = jwtDecode(cookies.token)
-    // Store userid from token to userid variable
-    const userid = decoded.user.id;
+    // State to hold userID
+    const [userid, setUserid] = useState('');
+
+    useEffect(() => {
+        if (cookies.token) {
+            try {
+                const decoded = jwtDecode(cookies.token);
+                setUserid(decoded.user.id);
+            } catch (error) {
+                console.error('Error decoding token:', error);
+            }
+        }
+    }, [cookies.token]);
 
     // Navigate to user profile on click
     function handleClickToProfile() {
@@ -36,7 +45,7 @@ function NavBar() {
         const confirmed = window.confirm(`Are you sure you want to 
             delete your account? This action cannot be undone.`);
 
-        if(confirmed)
+        if(confirmed && userid)
             try {
                 await axios({
                     method: "DELETE",
@@ -61,7 +70,7 @@ function NavBar() {
             <li className="spacer"></li>
             {isAuth ? (
                 <>
-                    <li className='logout'><button onClick={handleClickToDelete}>Profile</button></li>
+                    <li className='Delete'><button onClick={handleClickToDelete}>Delete Profile</button></li>
                     <li className='logout'><button onClick={handleClickToProfile}>Profile</button></li>
                     <li className='logout'><button onClick={handleClick}>Log Out</button></li>
                 </>
